Feature/combine setting activation state and user confirmable value (#28)
The purpose of this PR is to combine two manipulations into one. Setting
the activation state and a user confirmable value of a metric at once.
The manipulation SetUserConfirmableValue is then redundant and not used
anymore. It can be removed.

# Checklist

The following aspects have been respected by the author of this pull
request, confirmed by both pull request assignee **and** reviewer:

* Changelog update (necessity checked and entry added or not added
respectively)
  * [x] Pull Request Assignee
  * [x] Reviewer
* README update (necessity checked and entry added or not added
respectively)
  * [x] Pull Request Assignee
  * [x] Reviewer

---------

Co-authored-by: Staeck, Philipp <philipp.staeck.contractor@draeger.com>