Added trigger any descriptor update manipulation (#50)
The TriggerDescriptorUpdate manipulation requires specifying the
descriptor that is to be updated. However, for most descriptor types,
there are devices that do not support updating them. Also, to the best
of my knowledge, there are no type of descriptor that exists in every
Mdib and can be always be updated.
Hence, when the goal is to trigger a DescriptorManipulationReport with
parts of type UPT, then the only way to find a descriptor in the Mdib,
which can be updated is trial and error. However, trial and error would
involve calling the TriggerDescriptorUpdate manipulation quite often,
which is burdensome when the manipulation has to be performed manually.

Hence, we introduce a new manipulation TriggerAnyDescriptorUpdate which
essentially does the same as TriggerDescriptorUpdate, but does not
require the caller to specify the descriptor. Rather, it chooses the
descriptor itself - and as long as the device has descriptos that can be
updated, it should of course choose one of them.

# Checklist

The following aspects have been respected by the author of this pull
request, confirmed by both pull request assignee **and** reviewer:

* Changelog update (necessity checked and entry added or not added
respectively)
  * [x] Pull Request Assignee
  * [x] Reviewer
* README update (necessity checked and entry added or not added
respectively)
  * [x] Pull Request Assignee
  * [x] Reviewer

---------

Co-authored-by: Malte Grebe-Lüth <104556762+belagertem@users.noreply.github.com>
Co-authored-by: Ilja Antipov <33315542+antipovi@users.noreply.github.com>